Add WebSocket Server for Streaming Updates (#105)
This PR adds support for live-streaming updates to relics, characters, and gacha-info from the game. This also includes live-streaming of pull-results (not included in the static export obviously) These updates are formatted as events and broadcast on a websocket (by default at ws://localhost:53113).

Currently this feature is gated behind the stream feature, and --stream CLI flag.

It also includes a new GachaFunds struct in the export, that details how many oneric shards / stellar jade the account has.
Materials are also added to the export, so you can obtain the count of things like Star Rail Special Passes, Level Up materials, etc.